Move reflection based converter and contract resolution logic to DefaultJsonTypeInfoResolver
